wip: move some fundamental functions into new separate file
Update nuthouse01_pmx_utils.py

more moving

reorder functions
