+++ mmd_scripting/scripts_for_gui/translate_source_bone.py
@@ -1,7 +1,7 @@
 import mmd_scripting.core.nuthouse01_core as core
 import mmd_scripting.core.nuthouse01_pmx_parser as pmxlib
 import mmd_scripting.core.nuthouse01_pmx_struct as pmxstruct
-from mmd_scripting.overall_cleanup.prune_unused_bones import insert_single_bone
+from mmd_scripting.core.nuthouse01_pmx_utils import insert_single_bone
 
 _SCRIPT_VERSION = "Script version:  khanghugo - 9/21/2020 - v5.02"
